todolist_core/users/requirements.txt
amqp                          5.3.1
asgiref                       3.8.1
attrs                         22.2.0
billiard                      4.2.1
build                         0.10.0
CacheControl                  0.12.11
celery                        5.4.0
certifi                       2022.12.7
cffi                          1.15.1
charset-normalizer            3.1.0
cleo                          2.0.1
click                         8.1.8
click-didyoumean              0.3.1
click-plugins                 1.1.1
click-repl                    0.3.0
coverage                      7.6.10
crashtest                     0.4.1
cryptography                  39.0.2
distlib                       0.3.6
Django                        5.1.5
django-redis                  5.4.0
djangorestframework           3.15.2
djangorestframework_simplejwt 5.4.0
drf-yasg                      1.21.8
dulwich                       0.21.3
filelock                      3.9.0
flower                        2.0.1
html5lib                      1.1
humanize                      4.11.0
idna                          3.4
importlib-metadata            6.0.0
inflection                    0.5.1
installer                     0.6.0
jaraco.classes                3.2.3
jeepney                       0.8.0
jsonschema                    4.17.3
keyring                       23.13.1
kombu                         5.4.2
lockfile                      0.12.2
more-itertools                9.1.0
msgpack                       1.0.5
packaging                     23.0
pexpect                       4.8.0
pip                           24.3.1
pipenv                        2023.2.18
pkginfo                       1.9.6
platformdirs                  2.6.2
poetry                        1.4.0
poetry-core                   1.5.1
poetry-plugin-export          1.3.0
postgres                      4.0
prometheus_client             0.21.1
prompt_toolkit                3.0.48
psycopg2-binary               2.9.10
psycopg2-pool                 1.2
ptyprocess                    0.7.0
pycparser                     2.21
PyJWT                         2.10.1
pyproject_hooks               1.0.0
pyrsistent                    0.19.3
python-dateutil               2.9.0.post0
python-decouple               3.8
pytz                          2024.2
PyYAML                        6.0.2
rapidfuzz                     2.15.1
redis                         5.2.1
requests                      2.28.2
requests-toolbelt             0.10.1
rest-framework-simplejwt      0.0.2
SecretStorage                 3.3.3
setuptools                    68.2.0
shellingham                   1.5.0.post1
six                           1.17.0
sql                           2022.4.0
sqlparse                      0.5.3
tomlkit                       0.11.6
tornado                       6.4.2
trove-classifiers             2023.3.9
tzdata                        2024.2
uritemplate                   4.1.1
urllib3                       1.26.15
vine                          5.1.0
virtualenv                    20.20.0
virtualenv-clone              0.5.7
wcwidth                       0.2.13
webencodings                  0.5.1
wheel                         0.41.2
zipp                          3.15.0
